# Generate OrderedOptions accessors on first access