### Summary

Changes `ActiveSupport::OrderedOptions` to generate accessor methods for
each key on first access.

For the given key `:foo` 3 methods are generated: `#foo`, `#foo!` and `#foo=`

This change removes the requirement for
`ActiveSupport::Configuration#compile_methods!` as these methods no longer
need to be explicitly compiled.

Fixes #42140 

### Other Information

Chief motivation for this is as a bug fix for the fact that any keys on `ActiveSupport::OrderedOptions` are masked by methods on `Hash` when a method with the same name exists on `Hash`

```ruby
oo = ActiveSupport::OrderedOptions.new
oo.count = 99
oo.count # => 1
```

As a side effect of this change, speed is greatly improved in read and write method invocations as `method_missing` is bypassed.

```ruby
# frozen_string_literal: true

require "active_support"
require "active_support/ordered_options"
require "benchmark/ips"

oo = ActiveSupport::OrderedOptions.new

oo.foo = 1
oo.bar = "two"

Benchmark.ips do |x|
  x.report("write") { oo.baz = "three" }
  x.report("read")  { oo.foo ; oo.bar ; oo.not_found }
  x.report("read!")  { oo.foo! ; oo.bar! }
end
```
Before:
```
               write      1.355M (± 2.3%) i/s -      6.838M in   5.050324s
                read    485.645k (± 0.8%) i/s -      2.445M in   5.034469s
               read!    497.287k (± 0.6%) i/s -      2.519M in   5.066502s
```
After:
```
              write      3.148M (± 2.0%) i/s -     15.869M in   5.042414s
                read      1.298M (± 1.6%) i/s -      6.513M in   5.018765s
               read!    943.126k (± 6.3%) i/s -      4.752M in   5.064580s
```