@pixeltrix 2021-05-18
Interestingly this change was suggested here but never implemented: https://github.com/rails/rails/pull/56#issuecomment-432564

I wonder whether this was because of thread safety issues? A similar situation is where templates are dynamically compiled and created as methods and there [we use a mutex][1] to protect against recompilation.

[1]: https://github.com/rails/rails/blob/f95c0b7/actionview/lib/action_view/template.rb#L244